test(plugins): split install scan coverage
